pyproject.toml
[project]
name = "telegram-radar"
version = "0.1.0"
description = "Personal tool to summarize Telegram channels"
requires-python = ">=3.13"
dependencies = [
    "telethon>=1.37",
    "python-telegram-bot>=21",
    "instructor>=1.7",
    "openai>=1.60",
    "apscheduler>=3.10,<4",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "loguru>=0.7",
    "cryptg>=0.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/telegram_radar"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.24",
]
